heightおよびweightモデル入力フォーム、integer → floatに変更

+++ db/migrate/20240311041643_create_heights.rb
@@ -1,7 +1,7 @@
 class CreateHeights < ActiveRecord::Migration[7.0]
   def change
     create_table :heights do |t|
-      t.integer :height, null: false
+      t.float :height, null: false
 
       t.timestamps
     end
