+++ db/migrate/20240311220506_create_weights.rb
@@ -1,7 +1,7 @@
 class CreateWeights < ActiveRecord::Migration[7.0]
   def change
     create_table :weights do |t|
-      t.integer :weight, null: false
+      t.float :weight, null: false
 
       t.timestamps
     end
